requirements.txt
aioice==0.10.1
aiortc==1.14.0
altair==5.5.0
attrs==25.4.0
av==16.0.1
awscli==1.42.68
blinker==1.9.0
boto3==1.40.68
botocore==1.40.68
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
docutils==0.19
filelock==3.20.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.7.1
idna==3.11
ifaddr==0.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.10.2
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.23
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
rsa==4.7.2
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
streamlit-audiorec==0.1.3
streamlit-webrtc==0.63.11
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
